dash board has completed
