+++ lib/core/constant/app_string.dart
@@ -1,0 +1,1 @@
+class AppString{}
