+++ lib/feature/Screen/HomeScreen.dart
@@ -1,60 +1,57 @@
 import 'dart:async';
-import 'dart:math';
 
-import 'package:d_luscious/Authenticate/LoginScreen.dart';
-import 'package:d_luscious/Recipes/Pannertikkakabab.dart';
-import 'package:d_luscious/Recipes/Tacos.dart';
-import 'package:d_luscious/Recipes/cucumberraita.dart';
-import 'package:d_luscious/Recipes/springrolls.dart';
-import 'package:d_luscious/Recipes/veghakkanoodles.dart';
-import 'package:d_luscious/Recipes/whitesaucepasta.dart';
-import 'package:d_luscious/Screen/CategoryScreen.dart';
-import 'package:d_luscious/Screen/Favorite.dart';
-import 'package:d_luscious/Screen/SearchScreen.dart';
-import 'package:d_luscious/data/menu_items.dart';
-import 'package:d_luscious/model/menu_item.dart';
-import 'package:d_luscious/search_model/model.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 
-class Home extends StatefulWidget {
-  const Home({
+import '../Authenticate/LoginScreen.dart';
+import '../Recipes/Pannertikkakabab.dart';
+import '../Recipes/Tacos.dart';
+import '../Recipes/cucumberraita.dart';
+import '../Recipes/springrolls.dart';
+import '../Recipes/veghakkanoodles.dart';
+import '../Recipes/whitesaucepasta.dart';
+import '../data/menu_items.dart';
+import '../model/menu_item.dart';
+import '../search_model/model.dart';
+
+class HomeScreenTab extends StatefulWidget {
+  const HomeScreenTab({
     Key? key,
   }) : super(key: key);
 
   @override
-  State<Home> createState() => _HomeState();
+  State<HomeScreenTab> createState() => _HomeScreenTabState();
 }
 
-class _HomeState extends State<Home> {
+class _HomeScreenTabState extends State<HomeScreenTab> {
   late final PageController pageController;
 
-  ScrollController _scrollController = ScrollController();
+  final ScrollController _scrollController = ScrollController();
   int pageNo = 0;
 
-  List<String> listImages = [
+  final List<String> listImages = [
     "https://media.istockphoto.com/id/1408284950/photo/gulab-jamun-with-nuts-served-in-a-dish-isolated-on-napkin-side-view-on-dark-background-indian.jpg?b=1&s=170667a&w=0&k=20&c=JZ5cOreQ3eVfg3nsujNiWTfflLw-M3BvHSygip-lalE=",
     "https://as1.ftcdn.net/v2/jpg/01/95/70/12/1000_F_195701243_4uXJILIWRX6B6GUy4t9HQZt6HomkZbzT.jpg",
     "https://t3.ftcdn.net/jpg/02/54/45/12/240_F_254451278_7AAYmhYkBEVQ9MYnZYIomTI5WdFCif1g.jpg",
     "https://images.unsplash.com/photo-1602833280958-1657662ccc58?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=464&q=80",
     "https://t4.ftcdn.net/jpg/03/08/40/45/240_F_308404590_cKgzFkCSG93RG6ycBbQknhpr0y7pwmzP.jpg"
   ];
-  Timer? carasouelTmer;
+  Timer? carasouelTimer;
 
   Timer getTimer() {
     return Timer.periodic(const Duration(seconds: 3), (timer) {
       if (pageNo == 4) {
         pageNo = 0;
       }
-      pageController.animateToPage(
-        pageNo,
-        duration: const Duration(seconds: 1),
-        curve: Curves.easeInOutCirc,
-      );
+      if (pageController.hasClients) {
+        pageController.animateToPage(
+          pageNo,
+          duration: const Duration(seconds: 1),
+          curve: Curves.easeInOutCirc,
+        );
+      }
       pageNo++;
     });
   }
@@ -62,7 +59,7 @@ class _HomeState extends State<Home> {
   @override
   void initState() {
     pageController = PageController(initialPage: 0, viewportFraction: 0.85);
-    carasouelTmer = getTimer();
+    carasouelTimer = getTimer();
     _scrollController.addListener(() {
       if (_scrollController.position.userScrollDirection ==
           ScrollDirection.reverse) {
@@ -84,6 +81,7 @@ class _HomeState extends State<Home> {
 
   bool showBtmAppBr = true;
   final FirebaseAuth _auth = FirebaseAuth.instance;
+
   //logout alertbox fuunction
   void showAlert(BuildContext context) {
     showAlertDialog(BuildContext context) {
@@ -138,234 +136,228 @@ class _HomeState extends State<Home> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: SafeArea(
-        child: SingleChildScrollView(
-          controller: _scrollController,
-          child: Column(
-            children: [
-              const SizedBox(
-                height: 10.0,
-              ),
-              Padding(
-                padding: const EdgeInsets.only(
-                    left: 21, right: 21, top: 18.8, bottom: 20),
-                child: ListTile(
-                  onTap: () {},
-                  selected: true,
-                  shape: const RoundedRectangleBorder(
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(16.0),
-                    ),
-                  ),
-                  selectedTileColor: Color(0xffF5A342),
-                  title: Text(
-                    "Welcome",
-                    style: Theme.of(context).textTheme.subtitle1!.merge(
-                          const TextStyle(
-                            fontWeight: FontWeight.w700,
-                            fontSize: 18.0,
-                          ),
-                        ),
-                  ),
-                  subtitle: Text(
-                    "A Greet welcome to D'luscious.",
-                    style: Theme.of(context).textTheme.subtitle2,
-                  ),
-                  trailing: PopupMenuButton<MenuItemabc>(
-                    onSelected: (item) => onSelected(context, item),
-                    itemBuilder: (context) => [
-                      ...MenuItems.itemsFirst.map(buildItem).toList(),
-                      PopupMenuDivider(),
-                      ...MenuItems.itemsSecond.map(buildItem).toList(),
-                    ],
+      body: SingleChildScrollView(
+        controller: _scrollController,
+        child: Column(
+          children: [
+            const SizedBox(
+              height: 10.0,
+            ),
+            Padding(
+              padding: const EdgeInsets.only(
+                  left: 21, right: 21, top: 18.8, bottom: 20),
+              child: ListTile(
+                onTap: () {},
+                selected: true,
+                shape: const RoundedRectangleBorder(
+                  borderRadius: BorderRadius.all(
+                    Radius.circular(16.0),
                   ),
                 ),
+                selectedTileColor: Color(0xffF5A342),
+                title: Text(
+                  "Welcome",
+                  style: Theme.of(context).textTheme.subtitle1!.merge(
+                        const TextStyle(
+                          fontWeight: FontWeight.w700,
+                          fontSize: 18.0,
+                        ),
+                      ),
+                ),
+                subtitle: Text(
+                  "A Greet welcome to D'luscious.",
+                  style: Theme.of(context).textTheme.subtitle2,
+                ),
+                trailing: PopupMenuButton<MenuItemabc>(
+                  onSelected: (item) => onSelected(context, item),
+                  itemBuilder: (context) => [
+                    ...MenuItems.itemsFirst.map(buildItem).toList(),
+                    PopupMenuDivider(),
+                    ...MenuItems.itemsSecond.map(buildItem).toList(),
+                  ],
+                ),
               ),
-              SizedBox(
-                height: 200,
-                child: PageView.builder(
-                  controller: pageController,
-                  onPageChanged: (index) {
-                    pageNo = index;
-                    setState(() {});
-                  },
-                  itemBuilder: (_, index) {
-                    return AnimatedBuilder(
-                      animation: pageController,
-                      builder: (ctx, child) {
-                        return child!;
+            ),
+            SizedBox(
+              height: 200,
+              child: PageView.builder(
+                controller: pageController,
+                onPageChanged: (index) {
+                  pageNo = index;
+                  setState(() {});
+                },
+                itemBuilder: (_, index) {
+                  return AnimatedBuilder(
+                    animation: pageController,
+                    builder: (ctx, child) {
+                      return child!;
+                    },
+                    child: GestureDetector(
+                      onTap: () {
+                        ScaffoldMessenger.of(context).showSnackBar(
+                          SnackBar(
+                            content: Text("Hello you tapped at ${index + 1} "),
+                          ),
+                        );
                       },
-                      child: GestureDetector(
-                        onTap: () {
-                          ScaffoldMessenger.of(context).showSnackBar(
-                            SnackBar(
-                              content:
-                                  Text("Hello you tapped at ${index + 1} "),
-                            ),
-                          );
-                        },
-                        onPanDown: (d) {
-                          carasouelTmer?.cancel();
-                          carasouelTmer = null;
-                        },
-                        onPanCancel: () {
-                          carasouelTmer = getTimer();
-                        },
-                        child: Padding(
-                          padding: EdgeInsets.symmetric(horizontal: 25),
-                          child: Container(
-                            //padding: const EdgeInsets.all(10.0),
-                            decoration: BoxDecoration(
-                              image: DecorationImage(
-                                  image: NetworkImage(listImages[index]),
-                                  fit: BoxFit.cover),
-                              borderRadius: BorderRadius.circular(5.0),
-                            ),
+                      onPanDown: (d) {
+                        carasouelTimer?.cancel();
+                        carasouelTimer = null;
+                      },
+                      onPanCancel: () {
+                        carasouelTimer = getTimer();
+                      },
+                      child: Padding(
+                        padding: EdgeInsets.symmetric(horizontal: 25),
+                        child: Container(
+                          //padding: const EdgeInsets.all(10.0),
+                          decoration: BoxDecoration(
+                            image: DecorationImage(
+                                image: NetworkImage(listImages[index]),
+                                fit: BoxFit.cover),
+                            borderRadius: BorderRadius.circular(5.0),
                           ),
                         ),
                       ),
-                    );
-                  },
-                  itemCount: 5,
-                ),
+                    ),
+                  );
+                },
+                itemCount: 5,
               ),
-              const SizedBox(
-                height: 12.0,
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: List.generate(
-                  5,
-                  (index) => GestureDetector(
-                    child: Container(
-                      margin: const EdgeInsets.all(2.0),
-                      child: Icon(
-                        Icons.circle,
-                        size: 12.0,
-                        color: pageNo == index
-                            ? Colors.indigoAccent
-                            : Colors.grey.shade300,
-                      ),
+            ),
+            const SizedBox(
+              height: 12.0,
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: List.generate(
+                5,
+                (index) => GestureDetector(
+                  child: Container(
+                    margin: const EdgeInsets.all(2.0),
+                    child: Icon(
+                      Icons.circle,
+                      size: 12.0,
+                      color: pageNo == index
+                          ? Colors.indigoAccent
+                          : Colors.grey.shade300,
                     ),
                   ),
                 ),
               ),
-              Container(
-                padding:
-                    EdgeInsets.only(left: 30, bottom: 1, right: 20, top: 10),
-                alignment: Alignment.topLeft,
-                child: Text("Time Saving Recipes",
-                    style: TextStyle(
-                        fontSize: 30,
-                        fontFamily: "chewy",
-                        color: Color(0xffF5A342))),
-              ),
-              const Padding(
-                padding: EdgeInsets.all(20.0),
-                child: GridB(),
-              ),
-              Container(
-                padding:
-                    EdgeInsets.only(left: 30, bottom: 1, right: 20, top: 10),
-                alignment: Alignment.topLeft,
-                child: Text("Recommended For you",
-                    style: TextStyle(
-                        fontSize: 30,
-                        fontFamily: "chewy",
-                        color: Color(0xffF5A342))),
-              ),
-              const Padding(
-                padding: EdgeInsets.all(24.0),
-                child: GridB1(),
-              ),
-              Container(
-                padding:
-                    EdgeInsets.only(left: 30, bottom: 1, right: 20, top: 10),
-                alignment: Alignment.topLeft,
-                child: Text("Trending",
-                    style: TextStyle(
-                        fontSize: 30,
-                        fontFamily: "chewy",
-                        color: Color(0xffF5A342))),
-              ),
-              const Padding(
-                padding: EdgeInsets.all(24.0),
-                child: GridB2(),
-              ),
-            ],
-          ),
+            ),
+            Container(
+              padding: EdgeInsets.only(left: 30, bottom: 1, right: 20, top: 10),
+              alignment: Alignment.topLeft,
+              child: Text("Time Saving Recipes",
+                  style: TextStyle(
+                      fontSize: 30,
+                      fontFamily: "chewy",
+                      color: Color(0xffF5A342))),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(20.0),
+              child: GridB(),
+            ),
+            Container(
+              padding: EdgeInsets.only(left: 30, bottom: 1, right: 20, top: 10),
+              alignment: Alignment.topLeft,
+              child: Text("Recommended For you",
+                  style: TextStyle(
+                      fontSize: 30,
+                      fontFamily: "chewy",
+                      color: Color(0xffF5A342))),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(24.0),
+              child: GridB1(),
+            ),
+            Container(
+              padding: EdgeInsets.only(left: 30, bottom: 1, right: 20, top: 10),
+              alignment: Alignment.topLeft,
+              child: Text("Trending",
+                  style: TextStyle(
+                      fontSize: 30,
+                      fontFamily: "chewy",
+                      color: Color(0xffF5A342))),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(24.0),
+              child: GridB2(),
+            ),
+          ],
         ),
       ),
-      floatingActionButtonLocation: showBtmAppBr
-          ? FloatingActionButtonLocation.centerDocked
-          : FloatingActionButtonLocation.centerFloat,
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          // Navigator.push(
-          //     context,
-          //     MaterialPageRoute(
-          //       builder: (context) => SearchPage(),
-          //     ));
-        },
-        child: const Icon(
-          Icons.search,
-        ),
-      ),
-      bottomNavigationBar: AnimatedContainer(
-        child: BottomAppBar(
-          notchMargin: 8.0,
-          shape: const CircularNotchedRectangle(),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: [
-              IconButton(
-                onPressed: () {},
-                icon: const Icon(
-                  Icons.home_outlined,
-                ),
-              ),
-              IconButton(
-                onPressed: () {
-                  Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => FavoriteList(),
-                      ));
-                },
-                icon: const Icon(
-                  CupertinoIcons.heart,
-                ),
-              ),
-              const SizedBox(
-                width: 50,
-              ),
-              IconButton(
-                onPressed: () {
-                  Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => CategoryScreen(),
-                      ));
-                },
-                icon: const Icon(
-                  CupertinoIcons.circle_grid_3x3_fill,
-                ),
-              ),
-              IconButton(
-                onPressed: () {},
-                icon: const Icon(
-                  CupertinoIcons.bell,
-                ),
-              ),
-            ],
-          ),
-        ),
-        duration: const Duration(
-          milliseconds: 800,
-        ),
-        curve: Curves.easeInOutSine,
-        height: showBtmAppBr ? 70 : 0,
-      ),
+      // floatingActionButtonLocation: showBtmAppBr
+      //     ? FloatingActionButtonLocation.centerDocked
+      //     : FloatingActionButtonLocation.centerFloat,
+      // floatingActionButton: FloatingActionButton(
+      //   onPressed: () {
+      //     // Navigator.push(
+      //     //     context,
+      //     //     MaterialPageRoute(
+      //     //       builder: (context) => SearchPage(),
+      //     //     ));
+      //   },
+      //   child: const Icon(
+      //     Icons.search,
+      //   ),
+      // ),
+      // bottomNavigationBar: AnimatedContainer(
+      //   child: BottomAppBar(
+      //     notchMargin: 8.0,
+      //     shape: const CircularNotchedRectangle(),
+      //     child: Row(
+      //       mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+      //       children: [
+      //         IconButton(
+      //           onPressed: () {},
+      //           icon: const Icon(
+      //             Icons.home_outlined,
+      //           ),
+      //         ),
+      //         IconButton(
+      //           onPressed: () {
+      //             Navigator.push(
+      //                 context,
+      //                 MaterialPageRoute(
+      //                   builder: (context) => FavoriteScreenTab(),
+      //                 ));
+      //           },
+      //           icon: const Icon(
+      //             CupertinoIcons.heart,
+      //           ),
+      //         ),
+      //         const SizedBox(
+      //           width: 50,
+      //         ),
+      //         IconButton(
+      //           onPressed: () {
+      //             Navigator.push(
+      //                 context,
+      //                 MaterialPageRoute(
+      //                   builder: (context) => CategoryScreenTab(),
+      //                 ));
+      //           },
+      //           icon: const Icon(
+      //             CupertinoIcons.circle_grid_3x3_fill,
+      //           ),
+      //         ),
+      //         IconButton(
+      //           onPressed: () {},
+      //           icon: const Icon(
+      //             CupertinoIcons.bell,
+      //           ),
+      //         ),
+      //       ],
+      //     ),
+      //   ),
+      //   duration: const Duration(
+      //     milliseconds: 800,
+      //   ),
+      //   curve: Curves.easeInOutSine,
+      //   height: showBtmAppBr ? 70 : 0,
+      // ),
     );
   }
 
@@ -520,7 +512,7 @@ class _GridBState extends State<GridB> {
         crossAxisCount: 2,
         crossAxisSpacing: 12.0,
         mainAxisSpacing: 12.0,
-        mainAxisExtent: 260,
+        mainAxisExtent: 200,
       ),
       itemCount: click_c.length,
       itemBuilder: (_, index) {
@@ -534,68 +526,52 @@ class _GridBState extends State<GridB> {
               borderRadius: BorderRadius.circular(
                 10.0,
               ),
+              border: Border.all(width: 0.5),
               color: Colors.white,
             ),
             child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 ClipRRect(
                   borderRadius: const BorderRadius.only(
-                    topLeft: Radius.circular(16.0),
-                    topRight: Radius.circular(16.0),
+                    topLeft: Radius.circular(10.0),
+                    topRight: Radius.circular(10.0),
                   ),
                   child: Image.network(
                     click_c[index].images,
-                    height: 170,
                     width: double.infinity,
                     fit: BoxFit.cover,
                   ),
                 ),
-                Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  children: [
+                    Flexible(
+                      child: Text(
                         click_c[index].title,
-                        style: Theme.of(context).textTheme.subtitle1!.merge(
+                        style: Theme.of(context).textTheme.titleMedium!.merge(
                               const TextStyle(
                                 fontWeight: FontWeight.w700,
                               ),
                             ),
                       ),
-                      // Text(
-                      //   "${gridMap.elementAt(index)['price']}",
-                      //   style: Theme.of(context).textTheme.subtitle2!.merge(
-                      //         TextStyle(
-                      //           fontWeight: FontWeight.w700,
-                      //           color: Colors.grey.shade500,
-                      //         ),
-                      //       ),
-                      // ),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          IconButton(
-                            onPressed: () {
-                              setState(() {
-                                listIsFavTimeSaving[index] =
-                                    !listIsFavTimeSaving[index];
-                              });
-                            },
-                            icon: listIsFavTimeSaving[index]
-                                ? Icon(
-                                    Icons.favorite,
-                                    color: Colors.red,
-                                  )
-                                : Icon(
-                                    Icons.favorite_border,
-                                  ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
+                    ),
+                    IconButton(
+                      onPressed: () {
+                        setState(() {
+                          listIsFavTimeSaving[index] =
+                              !listIsFavTimeSaving[index];
+                        });
+                      },
+                      icon: listIsFavTimeSaving[index]
+                          ? Icon(
+                              Icons.favorite,
+                              color: Colors.red,
+                            )
+                          : Icon(
+                              Icons.favorite_border,
+                            ),
+                    ),
+                  ],
                 ),
               ],
             ),
