+++ lib/feature/model/a_network.dart
@@ -1,8 +1,9 @@
 import 'dart:convert';
+
 import 'package:flutter/foundation.dart';
+import 'package:http/http.dart' as http;
 
 import '../model/appitizermodel.dart';
-import 'package:http/http.dart' as http;
 
 class NetworkRequest {
   static const String url =
